requirements.txt
beautifulsoup4
geojsonio
geopandas
shapely
